+++ kompendium-auth/src/test/kotlin/org/leafygreens/kompendium/auth/util/TestData.kt
@@ -4,11 +4,11 @@ import java.io.File
 
 object TestData {
   object AuthConfigName {
-    val Basic = "basic"
-    val JWT = "jwt"
+    const val Basic = "basic"
+    const val JWT = "jwt"
   }
 
-  val getRoutePath = "/test"
+  const val getRoutePath = "/test"
 
   fun getFileSnapshot(fileName: String): String {
     val snapshotPath = "src/test/resources"